+++ src/include/sepuence.h
@@ -46,7 +46,7 @@ public:
 		{
 			NodeIncrement();
 
-			if (m_RanningIndex >= m_ChildeNode.size())
+			if (m_RanningIndex >= static_cast<int>(m_ChildeNode.size()))
 			{
 				m_NodeResult = NodeInfo::NodeResult::Re_SUCCESS;
 				Exit();
